add exception handler to delete method and create a unit test

+++ src/main/kotlin/com/api/v1/onboarding/service/implementation/ArtistService.kt
@@ -49,6 +49,11 @@ class ArtistService (
     override fun deleteArtistById(id: Int) {
         val artist = getById(id)
 
+        if(artist.status == ArtistStatus.USER_INACTIVE) throw BadRequestException(
+            DefaultExceptionResponse.ML203.message.format(id),
+            DefaultExceptionResponse.ML203.code
+        )
+
         portfolioService.deleteByArtist(artist)
 
         artist.status = ArtistStatus.USER_INACTIVE
